% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHeatmap.R
\name{buildHeatmap}
\alias{buildHeatmap}
\title{Function to combine the plots}
\usage{
buildHeatmap(heatmapClass, PRINT = TRUE)
}
\arguments{
\item{heatmapClass}{object of class heatmap (required)}

\item{PRINT}{TRUE/FALSE should the plot be displayed (default: TRUE)}
}
\description{
Build all the plots together
}
\examples{
heatmapClass <- plotHeatmap(heatmapClass, high = "red", mid = "white", low = "blue", midpoint = 0)
}
\keyword{heatmap}
\keyword{plot}
