% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{eSet2}
\alias{eSet2}
\title{Example dataset #2
TCGA breast cancer RNA-seq expression set (mad filtered 1000 genes), log2 transformed}
\format{expression set object}
\usage{
data(eSet2)
}
\description{
Example dataset #2
TCGA breast cancer RNA-seq expression set (mad filtered 1000 genes), log2 transformed
}

