% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{eSet3}
\alias{eSet3}
\title{Example dataset #3
TCGA breast cancer microarray expression set (mad filtered 1000 genes), gene median centered}
\format{expression set object}
\usage{
data(eSet3)
}
\description{
Example dataset #3
TCGA breast cancer microarray expression set (mad filtered 1000 genes), gene median centered
}

