% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateColDendro.R
\name{generateColDendro}
\alias{generateColDendro}
\title{Function for col dendro}
\usage{
generateColDendro(heatmapClass)
}
\arguments{
\item{heatmapClass}{object of class heatmap (required)}
}
\description{
uses 1 - spearman correlation to generate a distance matrix and then performs hclust to cluster the columns.
}
\examples{
heatmapClass <- generateColDendro(heatmapClass)
}
\keyword{cluster}
\keyword{rows}
