% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRowDendro.R
\name{generateRowDendro}
\alias{generateRowDendro}
\title{Function for row dendro}
\usage{
generateRowDendro(heatmapClass)
}
\arguments{
\item{heatmapClass}{object of class heatmap (required)}
}
\description{
uses 1 - pearson correlation to generate a distance matrix and then performs hclust to cluster the columns.
}
\examples{
heatmapClass <- generateRowDendro(heatmapClass)
}
\keyword{cluster}
\keyword{columns}
