% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Draw a simple heatmap based on ggplot2}
\usage{
ggheatmap(data, midpoint, scale = FALSE, show.num = FALSE, size = 4)
}
\arguments{
\item{data}{a data.frame with the first collumn as the the sample names usually transformed
from the rownames.}

\item{midpoint}{a numeric indicated the middle value of the heatmap legend.}

\item{scale}{Boolean value to indicate if scaled using the scale function.}

\item{show.num}{Boolean value to indicate if show numbers}

\item{size}{a numeric to indicate the size of numbers}
}
\value{
a heatmap
}
\description{
Draw a simple heatmap based on ggplot2
}
\examples{
data <- rowname_to_col(iris)
data <- data[,-6]
data$samples <- factor(data$samples, levels = data$samples)
ggheatmap(data = data, midpoint = 2, scale = TRUE)
}
\author{
Zhonghui Gai
}
