% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap_anno_sample.R
\name{ggheatmap_anno_sample}
\alias{ggheatmap_anno_sample}
\title{Draw a heatmap of the grouped samples, and add group information to the result of ggheatmap}
\usage{
ggheatmap_anno_sample(data)
}
\arguments{
\item{data}{a data.frame containg the samples information and the grouping information}
}
\value{
a ggplot2 type of heatmap containing only one row
}
\description{
Draw a heatmap of the grouped samples, and add group information to the result of ggheatmap
}
\examples{
data <- rowname_to_col(iris)
data <- data[,-6]
data$samples <- factor(data$samples, levels = data$samples)
p <- ggheatmap(data = data, midpoint = 2, scale = TRUE)
anno.samples <- data.frame(group = iris$Species,
samples = data$samples)
p.sample <- ggheatmap_anno_sample(data = anno.samples) + ggsci::scale_fill_jama()
p \%>\% aplot::insert_top(p.sample, height = 0.01)
}
\author{
Zhonghui Gai
}
