% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap_anno_variable.R
\name{ggheatmap_anno_variable}
\alias{ggheatmap_anno_variable}
\title{Draw a heatmap of the grouped variables, and add grouped information to the result of ggheatmap}
\usage{
ggheatmap_anno_variable(data)
}
\arguments{
\item{data}{a data.frame containing the variables information and the grouping information}
}
\value{
a ggplot2 type of heatmap containing only one collumn
}
\description{
Draw a heatmap of the grouped variables, and add grouped information to the result of ggheatmap
}
\examples{
data <- rowname_to_col(iris)
data <- data[,-6]
data$samples <- factor(data$samples, levels = data$samples)
p <- ggheatmap(data = data, midpoint = 2, scale = TRUE)
anno.variables <- data.frame(variables = colnames(iris)[1:4],
group = c(rep("grp1", 2), rep("grp2", 2)))

p.variable <- ggheatmap_anno_variable(data = anno.variables) + ggsci::scale_fill_lancet()
p \%>\% aplot::insert_left(p.variable, width = 0.03)
}
\author{
Zhonghui Gai
}
