% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap_cor1.R
\name{ggheatmap_cor1}
\alias{ggheatmap_cor1}
\title{Correlation heatmap of one data frame, Only the lower triangle is shown}
\usage{
ggheatmap_cor1(
  data,
  clust = TRUE,
  both = FALSE,
  method = "spearman",
  limits = c(-1, 1)
)
}
\arguments{
\item{data}{the numeric matrix}

\item{clust}{logic value to indicate whether cluster or not}

\item{both}{Logic value to show both the coefficient and the p values}

\item{method}{the method of correlation one the "pearson" and "spearman"}

\item{limits}{the color range, default value is c(-1, 1)}
}
\value{
a ggplot object
}
\description{
Correlation heatmap of one data frame, Only the lower triangle is shown
}
\examples{
ggheatmap_cor1(data = iris[, -5])
}
\author{
Zhonghui Gai
}
