% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap_cor2.R
\name{ggheatmap_cor2}
\alias{ggheatmap_cor2}
\title{Correlation heatmap of two matrix}
\usage{
ggheatmap_cor2(
  data1,
  data2,
  clust = TRUE,
  both = FALSE,
  method = "spearman",
  limits = c(-1, 1)
)
}
\arguments{
\item{data1}{the numeric matrix}

\item{data2}{the numeric matrix}

\item{clust}{logic value to indicate whether cluster or not}

\item{both}{Logic value to show both the coefficient and the p values}

\item{method}{the method of correlation one the "pearson" and "spearman"}

\item{limits}{the color range, default value is c(-1, 1)}
}
\value{

}
\description{
Correlation heatmap of two matrix
}
\examples{
ggheatmap_cor2(iris[, -5], iris[, -5])
}
\author{
Zhonghui Gai
}
