% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap_plotlist.R
\name{ggheatmap_plotlist}
\alias{ggheatmap_plotlist}
\title{Show ggheatmap Plot List}
\usage{
ggheatmap_plotlist(ggheatmap)
}
\arguments{
\item{ggheatmap}{heatmap, the result of ggheatmap}
}
\value{
plotlist
}
\description{
Display the basic elements of the ggheatmap
}
