% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap_test.R
\docType{data}
\name{ggheatmap_test}
\alias{ggheatmap_test}
\title{Sample data for ggheatmap}
\description{
This data is simulated data, not real data,
including 12 samples, 50 simulated gene expression data, and 3 annotation data.
}
\keyword{datasets}
