% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap_theme.R
\name{ggheatmap_theme}
\alias{ggheatmap_theme}
\title{The Theme of ggheatmap}
\usage{
ggheatmap_theme(ggheatmap, plotlist, theme)
}
\arguments{
\item{ggheatmap}{the result of ggheatmap}

\item{plotlist}{integer,the plotlist in ggheatmap. Use \link{ggheatmap_plotlist}}

\item{theme}{list, the theme of plotlist. More detail can see \link{theme}}
}
\value{
ggheatmap
}
\description{
Set the styles of the drawing elements of each component of the
ggheatmap except cluster tree.
}
