% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapClass.R
\docType{class}
\name{heatmapClass-class}
\alias{heatmapClass-class}
\alias{heatmapClass}
\title{Class for containing the heatmap and additional objects}
\arguments{
\item{RawData}{Matrix of gene expression (required)}

\item{ProcessedData}{Matrix of gene expression after processing (generated)}

\item{Plots}{list of all the plots generated (generated)}

\item{PlotData}{list of all data used to generate the plots. (generated)}

\item{ColInfo}{Data frame of info for column sidebars (optional)}

\item{RowData}{Data frame of info for row sidebars (optional)}

\item{Information}{list of various aspects of the data}
}
\description{
Class for containing the heatmap and additional objects
}
\examples{
heatmapClass <- heatmapClass(RawData = expressionMatrix)
}
\keyword{Mutation}
\keyword{Waterfall}
