% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap2.R
\name{pheatmap2}
\alias{pheatmap2}
\title{A swap function of pheatmap, the inputs can be one or two matix}
\usage{
pheatmap2(
  data1,
  data2 = NULL,
  cluster = FALSE,
  display_numbers = TRUE,
  cellwidth = 15,
  cellheight = 15
)
}
\arguments{
\item{data1}{a numeric matrix}

\item{data2}{a numeric matrix}

\item{cluster}{logic value to indicate whether to cluster or not}

\item{display_numbers}{logic value}

\item{cellwidth}{cellwidth}

\item{cellheight}{cellheight}
}
\value{

}
\description{
A swap function of pheatmap, the inputs can be one or two matix
}
\examples{
pheatmap2(iris[, -5])
pheatmap2(iris[c(1:10, 60:70, 100:110), -5])
pheatmap2(iris[, -5], cluster = T)
pheatmap2(iris[, -5], iris[, -5])
pheatmap2(iris[c(1:10, 60:70, 100:110), -5], display_numbers = T)
pheatmap2(t(iris[c(1:10, 60:70, 100:110), -5]), display_numbers = T, cellwidth = 25, cellheight = 25)
}
\author{
Zhonghui Gai
}
