% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSideBar.R
\name{plotColSideBar}
\alias{plotColSideBar}
\title{Function to run pipeline}
\usage{
plotColSideBar(heatmapClass, SampleData)
}
\arguments{
\item{heatmapClass}{object of class heatmap (required)}
}
\description{
Runs the entire pipeline to generate a heatmap
}
\examples{
heatmapClass <- plotHeatmap(heatmapClass, scale="row", Rowv = TRUE, Colv = TRUE)
}
\keyword{bars}
\keyword{colour}
\keyword{plot}
\keyword{side}
