% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendro.R
\name{plotDendro}
\alias{plotDendro}
\title{function to plot the dendrograms}
\usage{
plotDendro(heatmapClass, row = TRUE, col = TRUE)
}
\arguments{
\item{heatmapClass}{object of class heatmap (required)}

\item{row}{(TRUE/FALSE) plot dendro for rows (default: TRUE)}

\item{col}{(TRUE/FALSE) plot dendro for columns (default: TRUE)}
}
\description{
Plots the dendrograms from the hclust data stored in @PlotData
}
\examples{
heatmapClass <- plotDendro(heatmapClass, row=TRUE, col=TRUE)
}
\keyword{dendro}
\keyword{plot}
