% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Function to plot heatmap}
\usage{
plotHeatmap(
  heatmapClass,
  high = "red",
  mid = "white",
  low = "blue",
  midpoint = 0
)
}
\arguments{
\item{heatmapClass}{object of class heatmap (required)}

\item{high}{colour for high scores (default: red)}

\item{mid}{colour for mid scores (default: white)}

\item{low}{colour for high scores (default: blue)}

\item{midpoint}{score for mid colour (default: 0 as data is z score by default)}
}
\description{
Plots the dendrograms from the hclust data stored in PlotData
}
\keyword{heatmap}
\keyword{plot}
