% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowname_to_col.R
\name{rowname_to_col}
\alias{rowname_to_col}
\title{Move the rownames of a data.frame to the first collum of the data.frame}
\usage{
rowname_to_col(data, rowname.rm = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{rowname.rm}{a boolean to choose to remove the rownames}
}
\value{
a data.frame
}
\description{
Move the rownames of a data.frame to the first collum of the data.frame
}
\examples{
rowname_to_col(iris)
}
\author{
Zhonghui Gai
}
