% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Generate a ggplot2 heatmap with row and column dendrograms}
\usage{
ggheatmap(dataMatrix, orderCol = T, orderRow = T, dendroLineSize = 0.5,
  fontSize = 20, colorScheme = "Spectral", scaleName = "value",
  distMethod = "euclidean", clustMethod = "complete")
}
\arguments{
\item{orderCol}{Reorder the columns (default=T)}

\item{orderRow}{Reorder the rows (default=T)}

\item{dendroLineSize}{Size of the dendrogram lines (default=0.5)}

\item{fontSize}{Font size (default=20)}

\item{scaleName}{Name of the colorscale (default="value")}

\item{distMethod}{Distance method (default="euclidean", see ?dist)}

\item{clustMethod}{Clustering method (default="complete", see ?hclust)}

\item{colorPalette}{Color palette (default="Spectral")}
}
\description{
Generate a ggplot2 heatmap with row and column dendrograms
}
\examples{
ggheatmap(mtcars)
}
