% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Draw a simple heatmap based on ggplot2}
\usage{
ggheatmap(data, midpoint, scale = FALSE)
}
\arguments{
\item{data}{a data.frame with the first collumn as the the sample names usually transformed
from the rownames.}

\item{midpoint}{a numeric indicated the middle value of the heatmap legend.}

\item{scale}{Boolean value to indicate if scaled using the scale function.}
}
\value{
a heatmap
}
\description{
Draw a simple heatmap based on ggplot2
}
\examples{
data <- rowname_to_col(iris)
data <- data[,-6]
data$samples <- factor(data$samples, levels = data$samples)
ggheatmap(data = data, midpoint = 2, scale = TRUE)
}
\author{
Zhonghui Gai
}
